requirements.txt
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
pygame==2.1.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
